feat(deck-settings): add save button to all tabs on sheet layout
DeckSaveButton owns its own save lifecycle (loading, validation, close)
via injected editor + close key. Aside uses it directly; design, study,
and index tabs render it conditionally on sheet. Removes onSave/is_saving
from the modal root and the validate() expose from the aside.
