+++ src/components/modals/deck-settings/deck-aside.vue
@@ -1,40 +1,13 @@
 <script setup lang="ts">
-import { inject, ref, watch } from 'vue'
+import { inject } from 'vue'
 import { useI18n } from 'vue-i18n'
 import UiInput from '@/components/ui-kit/input.vue'
 import UiTextarea from '@/components/ui-kit/textarea.vue'
-import UiButton from '@/components/ui-kit/button.vue'
 import { deckEditorKey } from '@/composables/deck-editor'
-import { emitSfx } from '@/sfx/bus'
-
-type DeckAsideProps = {
-  loading?: boolean
-}
-
-const { loading } = defineProps<DeckAsideProps>()
+import DeckSaveButton from './deck-save-button.vue'
 
 const { t } = useI18n()
-const { settings, is_dirty } = inject(deckEditorKey)!
-
-const title_error = ref<string>()
-
-const emit = defineEmits<{ save: [] }>()
-
-/** Returns true if valid; sets error state and returns false otherwise. */
-function validate(): boolean {
-  if (settings.title?.trim()) return true
-  title_error.value = t('deck.create-modal.title-required')
-  return false
-}
-
-defineExpose({ validate })
-
-watch(
-  () => settings.title,
-  () => {
-    title_error.value = undefined
-  }
-)
+const { settings } = inject(deckEditorKey)!
 </script>
 
 <template>
@@ -45,7 +18,6 @@ watch(
     <div data-testid="deck-aside__inputs" class="flex flex-col gap-2">
       <ui-input
         :placeholder="t('deck.title-placeholder')"
-        :error="title_error"
         text-align="center"
         size="lg"
         v-model:value="settings.title"
@@ -58,19 +30,6 @@ watch(
       />
     </div>
 
-    <ui-button
-      data-testid="deck-aside__save-button"
-      data-theme="blue-500"
-      data-theme-dark="blue-650"
-      size="lg"
-      full-width
-      :loading="loading"
-      :disabled="!is_dirty"
-      :sfx="{ click: 'ui.snappy_button_2' }"
-      click-when-disabled
-      @click="is_dirty ? emit('save') : emitSfx('ui.digi_powerdown')"
-    >
-      {{ t('deck.settings-modal.submit-edit') }}
-    </ui-button>
+    <deck-save-button />
   </aside>
 </template>
