+++ src/components/modals/deck-settings/layout.ts
@@ -4,3 +4,6 @@ export type DeckSettingsLayout = 'sheet' | 'tablet' | 'desktop'
 
 export const deckSettingsLayoutKey: InjectionKey<ComputedRef<DeckSettingsLayout>> =
   Symbol('deck-settings-layout')
+
+export type DeckSettingsClose = (response?: boolean) => void
+export const deckSettingsCloseKey: InjectionKey<DeckSettingsClose> = Symbol('deck-settings-close')
